Switch to double from float

+++ include/vec3.h
@@ -1,28 +1,30 @@
+#include <iostream>
+
 #ifndef VEC3_HPP
 #define VEC3_HPP
 
 class vec3 {
     public:
         vec3() = default;
-        vec3(float x, float y, float z): e_{x, y, z} {};
+        vec3(double x, double y, double z): e_{x, y, z} {};
         vec3& operator+=(const vec3& other);
-        vec3& operator*=(const float c);
-        vec3& operator/=(const float c);
-        float x() const;
-        float y() const;
-        float z() const;
-        float length() const;
-        float length_squared() const;
+        vec3& operator*=(const double c);
+        vec3& operator/=(const double c);
+        double x() const;
+        double y() const;
+        double z() const;
+        double length() const;
+        double length_squared() const;
     private:
-        float e_[3];
+        double e_[3];
 };
 
 std::ostream& operator<<(std::ostream& os, const vec3& v);
 vec3 operator+(const vec3& a, const vec3& b);
 vec3 operator-(const vec3& a, const vec3& b);
-vec3 operator*(const vec3& a, float c);
-vec3 operator/(const vec3& a, float c);
-float dot(const vec3& a, const vec3& b);
+vec3 operator*(const vec3& a, double c);
+vec3 operator/(const vec3& a, double c);
+double dot(const vec3& a, const vec3& b);
 vec3 cross(const vec3& a, const vec3& b);
 vec3 unit_vector(vec3 v);
